+++ app/cells/visualization_components/specialities_visualization/specialities_visualization.js
@@ -15,7 +15,7 @@ ATSB.Components['visualization-components/specialities-visualization'] = Backbon
     var publicPercentage = []
     var privatePercentage = []
     totalSpecialities.forEach(function(item, index) {
-      publicPercentage.push(publicSpecialities[index][1])
+      publicPercentage.push(-publicSpecialities[index][1])
       privatePercentage.push(privateSpecialities[index][1])
     })
     var mainLabels = totalSpecialities.map(function(item){return item[0].toLowerCase()})
@@ -25,31 +25,39 @@ ATSB.Components['visualization-components/specialities-visualization'] = Backbon
       datasets: [{
           data: publicPercentage,
           label: 'Público',
-          backgroundColor: "#3fa6c922",
+          backgroundColor: "#3fa6c9bb",
+          hoverBackgroundColor: "#3fa6c9",
           borderColor: "#3fa6c9",
-          pointBorderColor: "#FFFFFF",
-          pointBackgroundColor: "#3fa6c9",
       }, {
           data: privatePercentage,
           label: 'Privado',
-          backgroundColor: "#FF3E2522",
-          borderColor: "#FF3E25",
-          pointBorderColor: "#FFFFFF",
-          pointBackgroundColor: "#FF3E2522",
+          backgroundColor: "#0B2F47bb",
+          hoverBackgroundColor: "#0B2F47",
+          borderColor: "#0B2F47",
       }]
     }
     this.chartOptions = {
-
-    };
+      tooltips: {
+        callbacks: {label: function(tooltipItem, data) {
+          var datasetIndex = tooltipItem.datasetIndex
+          var label = data.datasets[datasetIndex].label
+          return Math.abs(tooltipItem.yLabel) + " Sedes con esta especialidad en el sector " + label.toLowerCase();
+        }}
+      },
+      scales: {
+        xAxes: [{stacked: true, gridLines: false, display: false }],
+        yAxes: [{stacked: true, scaleOverride: true,
+          ticks : {callback: function(value) {return Math.abs(value)}},
+        }]
+      }
+    }
   },
   initChart: function() {
-    var ctx = this.$el.find("#myChart")
+    var ctx = this.$el.find("#speciality_chart")
     var myRadarChart = new Chart(ctx, {
-      type: 'radar',
+      type: 'bar',
       data: this.chartData,
       options: this.chartOptions
-    });
-
-
+    })
   },
 })